requirements.txt
python-frontmatter
markdown
Jinja2
beautifulsoup4